feat: enhance questionnaire functionality and add utility functions for question processing
- Updated `streamQuestion` to accept an optional AbortSignal for better request management.
- Introduced `psychologyQuestionUtils.ts` with functions to detect question types, extract options, derive Likert descriptors, compute profiles from responses, and classify investors.
