+++ package.json
@@ -19,12 +19,14 @@
     "axios": "^1.13.2",
     "chart.js": "^4.5.1",
     "cheerio": "^1.1.2",
-    "eslint-config-next": "^16.0.7",
+    "eslint-config-next": "^16.0.8",
     "framer-motion": "^12.23.25",
-    "gsap": "^3.13.0",
+    "gsap": "^3.14.1",
+    "html2canvas": "^1.4.1",
+    "jspdf": "^3.0.4",
     "lightweight-charts": "^5.0.9",
     "mssql": "^12.2.0",
-    "next": "^16.0.7",
+    "next": "^16.0.8",
     "react": "^19.2.1",
     "react-chartjs-2": "^5.3.1",
     "react-dom": "^19.2.1",
@@ -37,8 +39,8 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.39.1",
-    "@next/eslint-plugin-next": "^16.0.7",
-    "@types/node": "^24.10.1",
+    "@next/eslint-plugin-next": "^16.0.8",
+    "@types/node": "^24.10.2",
     "@types/react": "^19.2.7",
     "@types/react-dom": "^19.2.3",
     "@types/react-sparklines": "^1.7.5",
@@ -50,7 +52,7 @@
     "postcss": "^8.5.6",
     "tailwindcss": "^4.1.17",
     "typescript": "^5.9.3",
-    "typescript-eslint": "^8.48.1"
+    "typescript-eslint": "^8.49.0"
   },
   "description": "一個現代化的金融市場分析平台，提供股票等市場的技術分析、基本面分析、AI 預測等功能。",
   "main": "next.config.js",
