+++ src/pages/psychology/index.tsx
@@ -1,571 +1,74 @@
-import React, { useState } from "react";
-import ReactMarkdown from "react-markdown";
-import remarkGfm from "remark-gfm";
+import React from "react";
 import {
   HeartIcon,
   UserIcon,
-  DocumentTextIcon,
   ExclamationTriangleIcon,
   FaceSmileIcon,
 } from "@heroicons/react/24/outline";
 import PageHeader from "@/components/Layout/PageHeader";
-import * as PsychologyService from "@/services/PsychologyService";
+import usePsychology from "@/hooks/usePsychology";
+import {
+  deriveLikertDescriptor as deriveLikertDescriptorUtil,
+  classifyInvestor as classifyInvestorUtil,
+} from "@/utils/psychologyQuestionUtils";
 import Footer from "@/components/Layout/Footer";
+import ProgressPanel from "@/components/Psychology/ProgressPanel";
+import QuestionCard from "@/components/Psychology/QuestionCard";
+import ResultsPanel from "@/components/Psychology/ResultsPanel";
 
 export default function QuestionnairePage(): React.ReactElement {
-  // ...existing simple states...
-  const [sessionId, setSessionId] = useState<string | null>(null);
-  const [question, setQuestion] = useState<string | null>(null);
-  const [streamingQuestion, setStreamingQuestion] = useState<string>("");
-  const [isStreamingQuestion, setIsStreamingQuestion] =
-    useState<boolean>(false);
-  const [questionNumber, setQuestionNumber] = useState<number>(0);
-  const [answer, setAnswer] = useState<string>("");
-  const [finished, setFinished] = useState<boolean>(false);
-  const [advice, setAdvice] = useState<string | null>(null);
-  const [loading, setLoading] = useState<boolean>(false);
-  const [error, setError] = useState<string | null>(null);
-
-  // 新增：題型與互動狀態
-  const [questionType, setQuestionType] = useState<"open" | "mc" | "likert">(
-    "open"
-  );
-  const [options, setOptions] = useState<string[]>([]);
-  const [selectedIndex, setSelectedIndex] = useState<number | null>(null);
-  const [likertValue, setLikertValue] = useState<number>(3);
-  const [responses, setResponses] = useState<
-    { question: string; answer: string; type: string; value?: number | null }[]
-  >([]);
+  const {
+    sessionId,
+    question,
+    streamingQuestion,
+    isStreamingQuestion,
+    questionNumber,
+    answer,
+    setAnswer,
+    finished,
+    advice,
+    loading,
+    error,
+    questionType,
+    options,
+    streamedOptions,
+    selectedIndex,
+    setSelectedIndex,
+    likertValue,
+    setLikertValue,
+    // responses,
+    serverProfile,
+    setError,
+    investorType,
+    startTest,
+    streamQuestion,
+    submitAnswer,
+    resetTest,
+    computeProfile,
+  } = usePsychology();
   // 新增：從後端取得的 profile 與 investor type
-  const [serverProfile, setServerProfile] = useState<{
-    risk: number;
-    stability: number;
-    confidence: number;
-    patience: number;
-    sensitivity: number;
-  } | null>(null);
-  const [investorType, setInvestorType] = useState<string | null>(null);
+  // serverProfile, investorType from hook
   // const [showServerProfile, setShowServerProfile] = useState<boolean>(false);
 
   // 題型偵測與選項解析
-  const detectQuestionType = (q: string) => {
-    if (!q) return "open";
-    if (/(1\s*到\s*5|1-5|1~5|1～5|likert|1[^\d]*5)/i.test(q)) return "likert";
-    const chunks = q
-      .split(/[\n;/；、]| \/ | \| /)
-      .map((s) => s.trim())
-      .filter(Boolean);
-    return chunks.length >= 2 && chunks.length <= 10 ? "mc" : "open";
-  };
+  // helper for showing likert descriptor in UI
+  const deriveLikertDescriptor = (questionText: string | null, value: number) =>
+    deriveLikertDescriptorUtil(questionText || "", value);
 
-  const extractOptions = (q: string) => {
-    if (!q) return [];
+  // keep streamedOptions from hook (streamedOptions variable is from usePsychology)
 
-    // 優先以 "/" 分割，否則以換行/分號/or/pipe 分割
-    let parts = q
-      .split(/\s*\/\s*/)
-      .map((s) => s.trim())
-      .filter(Boolean);
-    if (parts.length < 2)
-      parts = q
-        .split(/\r?\n|;|；|、|\s+or\s+|\|/i)
-        .map((s) => s.trim())
-        .filter(Boolean);
-    if (parts.length < 2) return [];
+  // hook handles submission/streaming logic
 
-    // 清理前綴（編號、標點等）
-    parts = parts.map((p) => p.replace(/^[A-Za-z0-9).:、\s-]+/, "").trim());
+  // 相關 UI 已移至 QuestionCard 元件
 
-    // 若第一段像題目（過長或含問號/冒號），嘗試從該段取出可能的選項或丟棄題目段
-    const first = parts[0];
-    const looksLikeQuestion = first.length > 80 || /[？：:。]/.test(first);
-    if (looksLikeQuestion) {
-      const lastPunct = Math.max(
-        first.lastIndexOf("？"),
-        first.lastIndexOf("?"),
-        first.lastIndexOf("："),
-        first.lastIndexOf(":"),
-        first.lastIndexOf("。"),
-        first.lastIndexOf(".")
-      );
-      let candidate = first;
-      if (lastPunct !== -1 && lastPunct < first.length - 1) {
-        candidate = first.slice(lastPunct + 1).trim();
-        parts[0] = candidate;
-      } else if (first.includes("/")) {
-        const seg = first
-          .split("/")
-          .map((s) => s.trim())
-          .filter(Boolean);
-        if (seg.length > 1) parts[0] = seg.pop()!.trim();
-      } else {
-        // 若後續段落看起來更像選項，捨棄第一段
-        if (parts.slice(1).some((p) => p.length <= 80)) parts = parts.slice(1);
-        else {
-          const lastSpace = first.lastIndexOf(" ");
-          if (lastSpace !== -1 && lastSpace < first.length - 1)
-            parts[0] = first.slice(lastSpace + 1).trim();
-        }
-      }
-    }
+  // computeProfile moved to hook
 
-    const opts = parts
-      .map((p) => p.replace(/^["'“”]+|["'“”]+$/g, "").trim())
-      .filter(Boolean);
-    return opts.length >= 2 ? opts : [];
-  };
+  // classification moved to util/hook
+  // use util classification
 
-  const prepareQuestionUI = (qText: string) => {
-    const t = detectQuestionType(qText);
-    setQuestionType(t as "open" | "mc" | "likert");
-    if (t === "mc") {
-      setOptions(extractOptions(qText));
-      setSelectedIndex(null);
-    } else {
-      setOptions([]);
-      setSelectedIndex(null);
-    }
-    if (t === "likert") setLikertValue(3);
-  };
+  // Radar chart 已拆出為 `components/Psychology/RadarChart.tsx`
 
-  const startTest = async () => {
-    setLoading(true);
-    setError(null);
-    try {
-      const data = await PsychologyService.startQuestionnaire();
-      setSessionId(data.session_id);
-      setFinished(false);
-      setAdvice(null);
-      setAnswer("");
-      setQuestionNumber(0);
-      setResponses([]);
-      setLoading(false);
-      await streamQuestion(data.session_id, 0);
-    } catch (e: unknown) {
-      setError(String(e));
-      setLoading(false);
-    }
-  };
-
-  const [streamedOptions, setStreamedOptions] = useState<string[]>([]);
-
-  const streamQuestion = async (sessionId: string, questionNum: number) => {
-    setIsStreamingQuestion(true);
-    setStreamingQuestion("");
-    setQuestion(null);
-    setStreamedOptions([]); // 新增：每次串流前清空
-    try {
-      const response = await PsychologyService.streamQuestion(
-        sessionId,
-        questionNum
-      );
-      if (!response.ok) throw new Error(`HTTP ${response.status}`);
-      const reader = response.body?.getReader();
-      const decoder = new TextDecoder();
-      if (!reader) throw new Error("無法建立串流讀取器");
-      let accumulatedText = "";
-      while (true) {
-        const { done, value } = await reader.read();
-        if (done) break;
-        const chunk = decoder.decode(value);
-        const lines = chunk.split("\n");
-        for (const line of lines) {
-          if (line.startsWith("data: ")) {
-            try {
-              const data = JSON.parse(line.slice(6));
-              if (data.done) {
-                setQuestion(accumulatedText);
-                prepareQuestionUI(accumulatedText);
-                setIsStreamingQuestion(false);
-                setStreamedOptions([]); // 最終以正式選項為主
-                return;
-              }
-              accumulatedText += data.text;
-              setStreamingQuestion(accumulatedText);
-
-              // 新增：即時判斷題型與選項
-              const type = detectQuestionType(accumulatedText);
-              if (type === "mc") {
-                const opts = extractOptions(accumulatedText);
-                setStreamedOptions(opts);
-              } else {
-                setStreamedOptions([]);
-              }
-            } catch {
-              // ignore
-            }
-          }
-        }
-      }
-    } catch (e: unknown) {
-      setError(`串流錯誤: ${String(e)}`);
-      setIsStreamingQuestion(false);
-      setStreamedOptions([]);
-    }
-  };
-
-  // 提交回答：依題型整理答案、儲存本地回應並呼叫 API
-  const submitAnswer = async () => {
-    if (!sessionId) return setError("尚未開始測驗");
-    setLoading(true);
-    setError(null);
-
-    let finalAnswer = answer;
-    if (questionType === "mc") {
-      if (selectedIndex === null) {
-        setError("請選擇一個選項");
-        setLoading(false);
-        return;
-      }
-      finalAnswer = options[selectedIndex] || "";
-    } else if (questionType === "likert") {
-      // 組合數值與語意描述，讓分析模型能理解「程度與語意」
-      const qText = question || streamingQuestion || "";
-      const descriptor = deriveLikertDescriptor(qText, likertValue);
-      finalAnswer = `${likertValue} — ${descriptor}`;
-    } else finalAnswer = answer;
-
-    try {
-      const data = await PsychologyService.answerQuestion(
-        sessionId,
-        finalAnswer
-      );
-
-      // 儲存本地回應（用於備援分析）
-      setResponses((prev) => [
-        ...prev,
-        {
-          question: question || streamingQuestion || "",
-          answer: finalAnswer,
-          type: questionType,
-          value:
-            questionType === "likert"
-              ? Number(likertValue)
-              : questionType === "mc"
-              ? selectedIndex
-              : null,
-        },
-      ]);
-
-      if (!data.has_next_question) {
-        // 從後端取得 advice 與 profile, investor_type
-        setFinished(true);
-        setAdvice(data.advice || null);
-        setServerProfile(data.profile || null);
-        setInvestorType(data.investor_type || null);
-        setQuestion(null);
-        setStreamingQuestion("");
-        setIsStreamingQuestion(false);
-        setLoading(false);
-      } else {
-        setAnswer("");
-        setQuestionNumber((qn) => qn + 1);
-        setLoading(false);
-        await streamQuestion(sessionId, questionNumber + 1);
-      }
-    } catch (e: unknown) {
-      setError(String(e));
-      setLoading(false);
-    }
-  };
-
-  // 新增：根據題目語意推論 Likert 描述詞（返回中文描述）
-  const deriveLikertDescriptor = (questionText: string, value: number) => {
-    const q = (questionText || "").toLowerCase();
-    // 常見壓力/情緒關鍵字
-    const stressKeywords = ["壓力", "焦慮", "緊張", "擔心", "煩躁", "壓力大"];
-    const agreementKeywords = ["認同", "同意", "贊同"];
-    const riskKeywords = ["風險", "風險承受", "冒險", "風險偏好"];
-
-    // 基本提示詞
-    const baseLabels = ["從不", "偶爾", "有時", "經常", "非常常"];
-    const label = baseLabels[Math.max(0, Math.min(4, value - 1))];
-
-    // 根據題目判斷語意
-    if (stressKeywords.some((k) => q.includes(k))) {
-      // 強調「壓力/焦慮」語意
-      const suffix = [
-        "不會感到壓力",
-        "偶爾感到壓力",
-        "有時感到壓力",
-        "經常感到壓力",
-        "非常常感到壓力",
-      ];
-      return suffix[Math.max(0, Math.min(4, value - 1))];
-    }
-
-    if (riskKeywords.some((k) => q.includes(k))) {
-      const suffix = ["非常保守", "偏保守", "中性", "偏冒險", "非常冒險"];
-      return `${label}（${suffix[Math.max(0, Math.min(4, value - 1))]}）`;
-    }
-
-    if (agreementKeywords.some((k) => q.includes(k))) {
-      const suffix = [
-        "非常不認同",
-        "不認同",
-        "中立/有保留",
-        "認同",
-        "非常認同",
-      ];
-      return `${label}（${suffix[Math.max(0, Math.min(4, value - 1))]}）`;
-    }
-
-    // 預設回傳 base label
-    return `${label}`;
-  };
-
-  // MC / Likert UI helpers
-  const renderOptions = () => (
-    <div className="space-y-3">
-      {options.map((opt, idx) => (
-        <button
-          key={idx}
-          onClick={() => setSelectedIndex(idx)}
-          className={`w-full text-left px-4 py-3 rounded-lg border transition ${
-            selectedIndex === idx
-              ? "bg-purple-600 text-white border-transparent"
-              : "bg-white text-gray-800 border-gray-200 hover:shadow-sm"
-          }`}
-        >
-          {opt}
-        </button>
-      ))}
-    </div>
-  );
-
-  const renderLikert = () => {
-    const labels = ["從不", "偶爾", "有時", "經常", "非常常"]; // 1..5 的提示詞
-    return (
-      <div className="flex flex-col items-center space-y-2">
-        <div className="text-sm text-gray-600 mb-1">
-          請依程度選擇（1 = 最低，5 = 最高）
-        </div>
-        <div className="flex items-end justify-center space-x-3">
-          {labels.map((lbl, idx) => {
-            const v = idx + 1;
-            return (
-              <div key={v} className="flex flex-col items-center">
-                <button
-                  onClick={() => setLikertValue(v)}
-                  aria-label={`Likert ${v} - ${lbl}`}
-                  className={`px-3 py-2 rounded-md border focus:outline-none transition ${
-                    likertValue === v
-                      ? "bg-purple-600 text-white border-transparent"
-                      : "bg-white text-gray-800 border-gray-200"
-                  }`}
-                >
-                  {v}
-                </button>
-                <div className="text-xs text-gray-500 mt-1">{lbl}</div>
-              </div>
-            );
-          })}
-        </div>
-        <div className="text-xs text-gray-500 mt-2">
-          左：程度低 — 右：程度高
-        </div>
-      </div>
-    );
-  };
-
-  // 互動分析：計算 profile
-  const computeProfile = () => {
-    let risk = 50,
-      stability = 50,
-      confidence = 50,
-      patience = 50,
-      sensitivity = 50;
-
-    for (const r of responses) {
-      if (r.type === "likert" && typeof r.value === "number") {
-        const v = r.value;
-        risk += (v - 3) * 8;
-        stability += (3 - v) * 6;
-        confidence += (v - 3) * 6;
-        patience += (v - 3) * 4;
-        sensitivity += (3 - v) * 6;
-      } else if (r.type === "mc") {
-        const text = (r.answer || "").toLowerCase();
-        if (/(加碼|買入|進場|冒險)/.test(text)) {
-          risk += 12;
-          confidence += 8;
-          sensitivity += 6;
-        } else if (/(賣出|逃離|恐慌|立刻賣出|減碼)/.test(text)) {
-          risk -= 12;
-          stability -= 8;
-          sensitivity += 10;
-        } else if (/(觀望|冷靜|等待|持有)/.test(text)) {
-          stability += 10;
-          patience += 8;
-          risk -= 4;
-        } else {
-          if (typeof r.value === "number") {
-            const posFactor = r.value / Math.max(1, options.length - 1) - 0.5;
-            risk += posFactor * 16;
-            confidence += posFactor * 10;
-            sensitivity -= posFactor * 6;
-          }
-        }
-      } else {
-        const len = (r.answer || "").length;
-        if (len > 80) {
-          confidence += 6;
-          patience += 4;
-        }
-      }
-    }
-
-    const clamp = (v: number) => Math.max(0, Math.min(100, Math.round(v)));
-    return {
-      risk: clamp(risk),
-      stability: clamp(stability),
-      confidence: clamp(confidence),
-      patience: clamp(patience),
-      sensitivity: clamp(sensitivity),
-    };
-  };
-
-  const classifyInvestor = (p: {
-    risk: number;
-    stability: number;
-    confidence: number;
-    patience: number;
-    sensitivity: number;
-  }) => {
-    if (p.risk > 60 && p.stability < 40) return "波動型（情緒受市場影響）";
-    if (p.risk > 60 && p.stability >= 40) return "探險型（高風險偏好）";
-    if (p.risk <= 40 && p.stability >= 60) return "冷靜型（理性決策）";
-    if (p.risk <= 40 && p.stability < 60) return "謹慎型（保守穩健）";
-    return "綜合型（中庸平衡）";
-  };
-
-  // RadarChart SVG (調整預設 size，由 220 提升到 300)
-  const RadarChart = ({
-    values,
-    size = 300, // 放大預設尺寸
-  }: {
-    values: {
-      risk: number;
-      stability: number;
-      confidence: number;
-      patience: number;
-      sensitivity: number;
-    };
-    size?: number;
-  }) => {
-    const labels = [
-      "風險承受度",
-      "情緒穩定度",
-      "決策信心",
-      "長期耐心",
-      "市場敏感度",
-    ];
-    const dims = Object.values(values);
-
-    // Add an explicit margin around the chart inside the SVG viewBox so labels
-    // are rendered within the SVG canvas (prevents parent clipping).
-    const margin = Math.max(0, Math.round(size * 0.2));
-    const total = size + margin * 2;
-    const cx = total / 2;
-    const cy = total / 2;
-    // radius uses the inner "size" area so visual proportions don't change
-    const r = size / 2 - 20;
-    const points = dims.map((v, i) => {
-      const angle = (Math.PI * 2 * i) / dims.length - Math.PI / 2;
-      const radius = (v / 100) * r;
-      return `${cx + radius * Math.cos(angle)},${
-        cy + radius * Math.sin(angle)
-      }`;
-    });
-    const outer = labels.map((_, i) => {
-      const angle = (Math.PI * 2 * i) / labels.length - Math.PI / 2;
-      return `${cx + r * Math.cos(angle)},${cy + r * Math.sin(angle)}`;
-    });
-
-    return (
-      <div
-        style={{
-          width: "100%",
-          maxWidth: size + margin * 2,
-          margin: "0 auto",
-          padding: 0,
-          boxSizing: "border-box",
-          overflow: "visible",
-        }}
-      >
-        <svg
-          width="100%"
-          height="auto"
-          viewBox={`0 0 ${total} ${total}`}
-          preserveAspectRatio="xMidYMid meet"
-          className="mx-auto block"
-        >
-          {[0.2, 0.4, 0.6, 0.8, 1].map((s, idx) => {
-            const poly = labels
-              .map((_, i) => {
-                const angle = (Math.PI * 2 * i) / labels.length - Math.PI / 2;
-                const radius = s * r;
-                return `${cx + radius * Math.cos(angle)},${
-                  cy + radius * Math.sin(angle)
-                }`;
-              })
-              .join(" ");
-            return (
-              <polygon
-                key={idx}
-                points={poly}
-                fill="none"
-                stroke="#E6E6E6"
-                strokeWidth={1}
-              />
-            );
-          })}
-          {outer.map((pt, i) => {
-            const [x, y] = pt.split(",").map(Number);
-            // scale label offset with total to give enough space from the
-            // outer vertex; use a slightly larger offset now that we have a
-            // viewBox margin available
-            const offset = Math.max(10, Math.round(total * 0.03));
-            const dx = x > cx ? offset : x < cx ? -offset : 0;
-            const anchor = x > cx ? "start" : x < cx ? "end" : "middle";
-            return (
-              <text
-                key={i}
-                x={x + dx}
-                y={y}
-                fontSize={Math.max(10, Math.round(size * 0.04))}
-                fill="#374151"
-                textAnchor={anchor}
-                dominantBaseline="middle"
-              >
-                {labels[i]}
-              </text>
-            );
-          })}
-          <polygon
-            points={points.join(" ")}
-            fill="rgba(124,58,237,0.15)"
-            stroke="#7C3AED"
-            strokeWidth={2}
-          />
-          {points.map((pt, i) => {
-            const [x, y] = pt.split(",").map(Number);
-            return (
-              <circle
-                key={i}
-                cx={x}
-                cy={y}
-                r={Math.max(3, size * 0.01)}
-                fill="#7C3AED"
-              />
-            );
-          })}
-        </svg>
-      </div>
-    );
-  };
-
-  // UI constants & JSX — 保留原有整體佈局，但已在題目回答區與完成畫面整合上方功能
+  // UI constants & JSX — 保留原有整體邏輯，但改為全新版面格局（左側進度/摘要，右側題目）
   const Icon = FaceSmileIcon;
   const Title = "投資心理檢測";
   const Subtitle = "基於心理學的投資性格分析";
@@ -577,6 +80,28 @@ export default function QuestionnairePage(): React.ReactElement {
   const panelSubtitle2 = "專屬建議";
 
   const profile = computeProfile();
+  const estimatedTotalQuestions = 8; // 用於進度條與預估時間（若後端有總題數可以改成動態）
+
+  // 若用戶嘗試離開頁面，提示未保存之進度
+  React.useEffect(() => {
+    if (typeof window === "undefined") return;
+    const handler = (e: BeforeUnloadEvent) => {
+      if (sessionId && !finished) {
+        e.preventDefault();
+        e.returnValue = "您有未完成的測驗，確定要離開嗎？";
+      }
+    };
+    window.addEventListener("beforeunload", handler);
+    return () => window.removeEventListener("beforeunload", handler);
+  }, [sessionId, finished]);
+
+  // resetTest handled via hook
+
+  const computedInvestorType =
+    investorType ||
+    (serverProfile
+      ? classifyInvestorUtil(serverProfile)
+      : classifyInvestorUtil(profile));
 
   return (
     <>
@@ -593,322 +118,137 @@ export default function QuestionnairePage(): React.ReactElement {
 
       <div className="min-h-screen bg-gray-50">
         <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-8">
-          {/* 開始卡片 - 保留 */}
+          {/* 新版開場與左側進度面板 */}
           {!sessionId && (
-            <div className="bg-white rounded-xl shadow-lg animate-slideIn">
-              <div className="p-8 text-center">
-                <div className="mb-8">
+            <div className="bg-white rounded-xl shadow-lg animate-slideIn p-8">
+              <div className="grid grid-cols-1 lg:grid-cols-3 gap-6">
+                <div className="lg:col-span-1 flex flex-col items-center justify-center">
                   <div className="inline-flex items-center justify-center w-24 h-24 bg-linear-to-r from-purple-500 to-indigo-500 rounded-full mb-6 shadow-xl">
                     <HeartIcon className="w-12 h-12 text-white" />
                   </div>
-                  <h2 className="text-2xl font-semibold text-gray-800 mb-4">
-                    準備開始您的心理測評
+                  <h2 className="text-2xl font-semibold text-gray-800 mb-2">
+                    準備開始
                   </h2>
-                  <p className="text-gray-600 max-w-md mx-auto">
-                    我們將透過一系列問題來評估您的投資心理特質，整個過程大約需要
-                    5-10 分鐘
+                  <p className="text-gray-600 text-center">
+                    心理壓力與投資決策分析，約 5-10 分鐘。
                   </p>
                 </div>
-
-                <button
-                  onClick={startTest}
-                  disabled={loading}
-                  className="bg-linear-to-r from-purple-600 to-indigo-600 hover:from-purple-700 hover:to-indigo-700 text-white px-8 py-4 text-lg font-semibold rounded-lg min-w-48 disabled:opacity-50 disabled:cursor-not-allowed transform hover:scale-105 transition-all duration-200 shadow-lg"
-                >
-                  {loading ? (
-                    <div className="flex items-center justify-center">
-                      <div className="animate-spin rounded-full h-5 w-5 border-b-2 border-white mr-3"></div>
-                      準備中...
+                <div className="lg:col-span-2">
+                  <div className="flex items-center justify-between mb-4">
+                    <div>
+                      <h3 className="text-lg font-semibold">
+                        心理壓力檢測流程
+                      </h3>
+                      <p className="text-sm text-gray-500">
+                        逐題回答後即可獲得個人化分析及建議
+                      </p>
                     </div>
-                  ) : (
-                    <>
-                      <UserIcon className="w-5 h-5 mr-2 inline" />
-                      開始測驗
-                    </>
-                  )}
-                </button>
-              </div>
-            </div>
-          )}
-
-          {/* 問題回答卡片 */}
-          {sessionId && !finished && (
-            <div className="bg-white rounded-xl shadow-lg animate-slideIn">
-              <div className="border-b border-gray-200 px-6 py-4 bg-linear-to-r from-purple-50 to-indigo-50">
-                <div className="flex items-center justify-between">
-                  <h2 className="text-xl font-semibold text-gray-800 flex items-center">
-                    <div className="w-8 h-8 bg-linear-to-r from-purple-500 to-indigo-500 rounded-full flex items-center justify-center mr-3">
-                      <DocumentTextIcon className="w-4 h-4 text-white" />
-                    </div>
-                    問卷題目
-                  </h2>
-                  <div className="flex items-center space-x-3">
-                    <div className="text-sm text-gray-500 bg-white px-3 py-1 rounded-full border">
-                      第 {questionNumber + 1} 題
-                    </div>
-                    {isStreamingQuestion && (
-                      <div className="flex items-center text-sm text-purple-600">
-                        <div className="animate-spin rounded-full h-4 w-4 border-b-2 border-purple-600 mr-2"></div>
-                        AI 正在思考...
-                      </div>
-                    )}
-                  </div>
-                </div>
-              </div>
-
-              <div className="p-6">
-                <div className="mb-8">
-                  <label className="block text-sm font-medium text-gray-700 mb-3">
-                    題目：
-                  </label>
-                  <div className="bg-linear-to-r from-purple-50 to-indigo-50 border-l-4 border-purple-500 p-6 rounded-lg relative overflow-hidden">
-                    {isStreamingQuestion ? (
-                      <div className="text-gray-800 text-lg leading-relaxed">
-                        <ReactMarkdown remarkPlugins={[remarkGfm]}>
-                          {streamingQuestion}
-                        </ReactMarkdown>
-                      </div>
-                    ) : (
-                      <div className="text-gray-800 text-lg leading-relaxed">
-                        <ReactMarkdown remarkPlugins={[remarkGfm]}>
-                          {question || ""}
-                        </ReactMarkdown>
-                      </div>
-                    )}
-                  </div>
-                </div>
-
-                <div className="mb-6">
-                  <label className="block text-sm font-medium text-gray-700 mb-3">
-                    您的回答：
-                  </label>
-
-                  {/* 串流時即顯示選項（streamedOptions），否則用正式 options */}
-                  {isStreamingQuestion && streamedOptions.length > 0 && (
-                    <div className="mb-4">
-                      <div className="text-xs text-gray-500 mb-2">
-                        （選項預覽，AI生成中）
-                      </div>
-                      <div className="space-y-3">
-                        {streamedOptions.map((opt, idx) => (
-                          <button
-                            key={idx}
-                            disabled
-                            className="w-full text-left px-4 py-3 rounded-lg border bg-gray-100 text-gray-400 border-gray-200"
-                          >
-                            {opt}
-                          </button>
-                        ))}
-                      </div>
-                    </div>
-                  )}
-
-                  {questionType === "mc" && !isStreamingQuestion && (
-                    <div className="mb-4">{renderOptions()}</div>
-                  )}
-                  {questionType === "likert" && !isStreamingQuestion && (
-                    <div className="mb-4">{renderLikert()}</div>
-                  )}
-                  {questionType === "open" && !isStreamingQuestion && (
-                    <textarea
-                      value={answer}
-                      onChange={(e) => setAnswer(e.target.value)}
-                      rows={6}
-                      className="w-full px-4 py-3 border border-gray-300 rounded-xl focus:ring-2 focus:ring-purple-500 focus:border-transparent transition-all duration-200 resize-none bg-white/80 backdrop-blur-sm shadow-sm hover:shadow-md"
-                      placeholder="請在此輸入您的想法和回答（中文）..."
-                      disabled={isStreamingQuestion}
-                    />
-                  )}
-
-                  <div className="mt-2 flex justify-between items-center">
                     <div className="text-sm text-gray-500">
-                      字數：
-                      {questionType === "open"
-                        ? answer.length
-                        : questionType === "likert"
-                        ? likertValue
-                        : selectedIndex !== null
-                        ? selectedIndex + 1
-                        : 0}
+                      預估用時：5 - 10 分鐘
                     </div>
-                    {isStreamingQuestion && (
-                      <div className="text-sm text-amber-600 flex items-center">
-                        <ExclamationTriangleIcon className="w-4 h-4 mr-1" />
-                        請等待問題生成完成
-                      </div>
-                    )}
                   </div>
-                </div>
-
-                <div className="flex justify-end">
-                  <button
-                    onClick={submitAnswer}
-                    disabled={
-                      loading ||
-                      (questionType === "open" && !answer.trim()) ||
-                      (questionType === "mc" && selectedIndex === null) ||
-                      (questionType === "likert" && !likertValue)
-                    }
-                    className="bg-linear-to-r from-purple-600 to-indigo-600 hover:from-purple-700 hover:to-indigo-700 text-white px-6 py-3 font-medium rounded-lg disabled:opacity-50 disabled:cursor-not-allowed transform hover:scale-105 transition-all duration-200 shadow-lg"
-                  >
-                    {loading ? (
-                      <div className="flex items-center">
-                        <div className="animate-spin rounded-full h-4 w-4 border-b-2 border-white mr-2"></div>
-                        提交中...
-                      </div>
-                    ) : (
-                      <>
-                        <span>送出回答</span>
-                      </>
-                    )}
-                  </button>
+                  <div className="rounded-lg border border-gray-100 bg-linear-to-r from-purple-50 to-indigo-50 p-4">
+                    <ol className="flex items-center justify-between">
+                      {["啟動", "問答", "結果"].map((s, idx) => {
+                        const activeStep = sessionId ? (finished ? 2 : 1) : 0;
+                        return (
+                          <li key={s} className="flex-1 text-center">
+                            <div
+                              className={`mx-auto w-9 h-9 flex items-center justify-center rounded-full ${
+                                idx === activeStep
+                                  ? "bg-purple-600 text-white"
+                                  : "bg-white text-gray-600"
+                              } border`}
+                            >
+                              {idx + 1}
+                            </div>
+                            <div className="mt-2 text-xs text-gray-600">
+                              {s}
+                            </div>
+                          </li>
+                        );
+                      })}
+                    </ol>
+                  </div>
+                  <div className="mt-6 flex justify-start">
+                    <button
+                      onClick={startTest}
+                      disabled={loading}
+                      className="bg-linear-to-r from-purple-600 to-indigo-600 hover:from-purple-700 hover:to-indigo-700 text-white px-6 py-3 text-lg font-semibold rounded-lg disabled:opacity-50 disabled:cursor-not-allowed transform hover:scale-105 transition-all duration-200 shadow"
+                    >
+                      {loading ? (
+                        <div className="flex items-center">
+                          <div className="animate-spin rounded-full h-4 w-4 border-b-2 border-white mr-2"></div>
+                          準備中...
+                        </div>
+                      ) : (
+                        <>
+                          <UserIcon className="w-5 h-5 mr-2 inline" />
+                          開始測驗
+                        </>
+                      )}
+                    </button>
+                  </div>
                 </div>
               </div>
             </div>
           )}
 
-          {/* 完成結果：AI 建議 + 互動分析雷達圖（重構版） */}
-          {finished && (
-            <section
-              aria-labelledby="result-heading"
-              className="bg-white rounded-xl shadow-lg animate-slideIn"
-            >
-              <header className="border-b border-gray-200 px-6 py-4 bg-linear-to-r from-green-50 to-emerald-50">
-                <div className="flex items-center">
-                  <div className="w-8 h-8 bg-linear-to-r from-green-500 to-emerald-500 rounded-full flex items-center justify-center mr-3">
-                    <svg
-                      className="w-4 h-4 text-white"
-                      fill="none"
-                      stroke="currentColor"
-                      viewBox="0 0 24 24"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth={2}
-                        d="M5 13l4 4L19 7"
-                      />
-                    </svg>
-                  </div>
-                  <h2
-                    id="result-heading"
-                    className="text-xl font-semibold text-gray-800"
-                  >
-                    測驗完成 — 個人化建議
-                  </h2>
-                </div>
-              </header>
-
-              <div className="p-6">
-                <div className="mb-6">
-                  <div className="bg-linear-to-r from-green-50 to-emerald-50 border-l-4 border-green-500 p-6 rounded-lg">
-                    <h3 className="text-lg font-semibold text-green-800 mb-3 flex items-center">
-                      您的理財心理分析報告
-                    </h3>
-
-                    <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-                      {/* 左：AI 建議 */}
-                      <aside
-                        className="bg-white rounded-lg p-4 shadow-sm"
-                        aria-label="AI 建議"
-                      >
-                        <h4 className="text-sm font-medium text-gray-700 mb-2">
-                          AI 建議
-                        </h4>
-                        <div className="text-gray-700 leading-relaxed font-sans text-lg">
-                          {advice ? (
-                            <pre className="whitespace-pre-wrap">{advice}</pre>
-                          ) : (
-                            <div className="text-gray-500">
-                              目前沒有建議內容。
-                            </div>
-                          )}
-                        </div>
-                      </aside>
-
-                      {/* 右：互動分析 + 雷達圖 */}
-                      <div
-                        className="bg-white rounded-lg p-4 shadow-sm text-center"
-                        aria-label="互動分析"
-                      >
-                        <h4 className="text-sm font-medium text-gray-700 mb-2">
-                          互動分析
-                        </h4>
-                        <div className="flex items-center justify-center">
-                          <div className="w-full max-w-lg">
-                            <RadarChart
-                              values={serverProfile || profile}
-                              size={340}
-                            />
-                          </div>
-                        </div>
-
-                        <div className="mt-4">
-                          <div className="text-sm text-gray-600">
-                            投資者類型
-                          </div>
-                          <div className="text-lg font-semibold text-gray-800">
-                            {investorType ||
-                              (serverProfile
-                                ? classifyInvestor(serverProfile)
-                                : classifyInvestor(profile))}
-                          </div>
-                        </div>
-
-                        {/* 顯示伺服器 profile（若有）或本地計算值 */}
-                        <div className="mt-4 text-left">
-                          <div className="text-xs text-gray-500 mb-2">
-                            指標（0-100）
-                          </div>
-                          <div className="grid grid-cols-2 gap-2 text-sm text-gray-700">
-                            {Object.entries(serverProfile || profile).map(
-                              ([k, v]) => (
-                                <div
-                                  key={k}
-                                  className="flex justify-between items-center bg-gray-50 rounded-md px-3 py-2"
-                                >
-                                  <div className="capitalize text-gray-600">
-                                    {k.replace(/([A-Z])/g, " $1")}
-                                  </div>
-                                  <div className="font-medium text-gray-800">
-                                    {v}
-                                  </div>
-                                </div>
-                              )
-                            )}
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-
-                <div className="flex justify-center">
-                  <button
-                    onClick={() => {
-                      setSessionId(null);
-                      setQuestion(null);
-                      setAnswer("");
-                      setFinished(false);
-                      setAdvice(null);
-                      setStreamingQuestion("");
-                      setIsStreamingQuestion(false);
-                      setQuestionNumber(0);
-                      setResponses([]);
-                      setQuestionType("open");
-                      setOptions([]);
-                      setSelectedIndex(null);
-                      setLikertValue(3);
-                      setServerProfile(null);
-                      setInvestorType(null);
+          {/* 問題回答卡片（新版：兩欄）*/}
+          {sessionId && !finished && (
+            <div className="bg-white rounded-xl shadow-lg animate-slideIn p-0 overflow-hidden">
+              <div className="grid grid-cols-1 md:grid-cols-4 gap-0">
+                {/* 左側欄：進度與輔助說明 (已拆成元件) */}
+                <ProgressPanel
+                  questionNumber={questionNumber}
+                  estimatedTotalQuestions={estimatedTotalQuestions}
+                  onCancel={() => {
+                    if (confirm("確定要取消測驗，所有進度將會遺失？"))
+                      resetTest();
+                  }}
+                  sessionId={sessionId}
+                  finished={finished}
+                />
+                {/* 右側大區塊：題目與回答 */}
+                <main className="md:col-span-3 p-6">
+                  <QuestionCard
+                    question={question}
+                    streamingQuestion={streamingQuestion}
+                    isStreamingQuestion={isStreamingQuestion}
+                    questionType={questionType}
+                    options={options}
+                    streamedOptions={streamedOptions}
+                    selectedIndex={selectedIndex}
+                    setSelectedIndex={(idx) => setSelectedIndex(idx)}
+                    likertValue={likertValue}
+                    setLikertValue={(v) => setLikertValue(v)}
+                    likertDescriptor={deriveLikertDescriptor(
+                      question || streamingQuestion,
+                      likertValue
+                    )}
+                    onRegenerate={() => {
+                      if (sessionId) streamQuestion(sessionId, questionNumber);
                     }}
-                    className="bg-linear-to-r from-gray-600 to-gray-700 hover:from-gray-700 hover:to-gray-800 text-white px-6 py-3 font-medium rounded-lg transform hover:scale-105 transition-all duration-200 shadow-lg"
-                  >
-                    重新測驗
-                  </button>
-                </div>
+                    answer={answer}
+                    setAnswer={(s) => setAnswer(s)}
+                    submitAnswer={submitAnswer}
+                    loading={loading}
+                    questionNumber={questionNumber}
+                  />
+                </main>
               </div>
-            </section>
+            </div>
+          )}
+
+          {/* 完成結果（新版卡片）：AI 建議 + 互動分析雷達圖 */}
+          {finished && (
+            <ResultsPanel
+              advice={advice}
+              serverProfile={serverProfile}
+              profile={profile}
+              investorType={computedInvestorType}
+              onReset={resetTest}
+            />
           )}
 
           {/* 載入指示器 */}
